projects/capstone/backend/requirements.txt
alembic==1.4.2
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.3.3
attrs==18.1.0
auth==0.5.3
Authlib==0.14.3
autopep8==1.4.4
Babel==2.8.0
beautifulsoup4==4.8.2
blinker==1.4
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
ConfigArgParse==1.0
coursera-dl==0.11.5
cryptography==2.9.2
distlib==0.3.0
dnspython==1.16.0
ecdsa==0.15
eventlet==0.25.2
falcon==2.0.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jose==1.0.0
keyring==21.1.0
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.20.0
monotonic==1.5
numpy==1.18.1
opencv-contrib-python==4.2.0.32
Pillow==7.0.0
pipenv==2018.11.26
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
pymongo==3.10.1
python-dateutil==2.6.0
python-editor==1.0.4
python-jose==3.1.0
pytz==2020.1
pywin32-ctypes==0.2.0
requests==2.22.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.16
urllib3==1.25.8
virtualenv==20.0.20
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.3.1
python-dotenv==0.20.0
